nftw, nftw64: add wrapper
Add a wrapper for nftw and ftw[1] calls (along with nftw64 and ftw64).

The call in brief: it accepts a path, which it walks. For every
entries it finds, it calls a user-specified callback function, and
passes some information about the entry to this callback.

The implementation saves the callback from the nftw call, and subtitutes
it with its own "fake_callback". When the real nftw calls the fake_callback,
it corrects the stat struct it received with information queried from pseudo.
Afterwards it calls the original callback and passes the now corrected
information to it.

Since nftw and nftw64 are so similar, the same codebase is used to
implement both (which is also fairly similar to their implementation in
glibc).

[1]: https://linux.die.net/man/3/nftw

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>

Rework nftw_wrapper_base to provide a 'pseudo_<name>' function for the
generated wrapper functions to call.  This is cleaner for debugging and
profiling in the future as the standard wrapper is now being used.

Reworded the commit message above to remove references to a chunk that
is no longer applicable.

Added error checking around chdir and fchdir function calls.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>