makewrappers: Introduce 'array' support
Allow for arrays to be specified in the declaration to avoid warnings about
type conversions from an array to a pointer.

Move three functions that glibc uses arrays for to the linux ports.  These
array based declarations might be valid elsewhere, but I don't really have
a way to test it.  Leave other ports 'as-is'.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>