Move ftw and ftw64 to calling ntfw and nftw64
The 4 functions are very similar to each other: nftw-nftw64 and ftw-ftw64 are
identical to their pair, except for the stat struct they use (stat vs stat64).

nftw is a "superset" of ftw: nftw is backwards compatible with ftw, but it
also accepts a number of extra flags to modify its behavior.

Since all the 4 functions are so similar, the same codebase is used to
implement all (which is also fairly similar to their implementation in
glibc).

[1]: https://linux.die.net/man/3/nftw

Re-implemented, but based on the ideas from Gyorgy Sarvari <skandigraun@gmail.com>

Signed-off-by: Mark Hatle <mark.hatle@amd.com>