pseudo_db.c: Fix xattr memory allocation
The memory allocated for oldmsg, as well as the memcpy was incorrectly
determined by a missing calculation (oldmsg) and incorrect variable usage
sizeof(msg), which is the size of a pointer, vs sizeof(*msg), which is
the size of the actual data structure.

Bug reported and a suggested fix sent by Daniele Romano, see:
https://bugzilla.yoctoproject.org/show_bug.cgi?id=15925

The suggested fix was changed a bit, as the normal sizeof usage in pseudo
is against the variable being used. Instead of sizeof(type),
sizeof(*variable) follows existing code practices.

We use sizeof(variable) instead of sizeof(type) in case the type of a
variable is changed in the future.  This was just the style of the code
when pseudo was written, which we intend to keep unless it proves
problematic.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>